# src/lib/api/generated/types/FileInfo.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type FileInfo = {
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type string | undefined
    */
    name?: string;
    /**
     * @type integer | undefined, int64
    */
    size?: number;
    /**
     * @type string | undefined
    */
    mimeType?: string;
    /**
     * @type string | undefined, date-time
    */
    createdAt?: string;
    /**
     * @type string | undefined
    */
    url?: string;
};