# src/lib/api/generated/types/HealthResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type HealthResponse = {
    /**
     * @type string | undefined
    */
    status?: string;
    /**
     * @type string | undefined
    */
    service?: string;
};